address the pr comments
